We have created Classes and functions
We have created  Game, Round, Pot, and GameDemo class to round out the structure of a game. game demo is to test it. right now we can print out the base features of any amount of bot players and one real main player. while not being able to play a real game yet